--- src/main/java/tests/TaxComputingTest.java
@@ -1,15 +1,0 @@
-package tests;
-import io.javalin.Javalin;
-import org.junit.Test;
-
-import java.util.logging.Handler;
-
-import static org.mockito.Mockito.mock;
-
-public class TaxComputingTest {
-    @Test
-    public void computeTaxTest(){
-
-    }
-
-}
