+++ src/test/java/zpi/sales/TaxComputingTest.java
@@ -1,0 +1,16 @@
+package zpi.sales;
+
+import io.javalin.Javalin;
+import org.junit.Test;
+
+import java.util.logging.Handler;
+
+import static org.mockito.Mockito.mock;
+
+public class TaxComputingTest {
+    @Test
+    public void computeTaxTest(){
+
+    }
+
+}
